requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
tenacity>=8.2.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0